offchain/requirements.txt
flask
python-dotenv
pycardano
